style: move components into shared folder
Since these are Single File Components, they don't need to be in their
own folder.
